backend/requirements.txt
Flask==2.3.3
Flask-JWT-Extended==4.5.2
Flask-SQLAlchemy==3.1.1
Flask-Limiter==3.5.0
Flask-Cors==4.0.0
SQLAlchemy==2.0.21
openai==0.28.1
python-dotenv==1.0.0
requests==2.31.0
bcrypt==4.0.1
email-validator==2.0.0
marshmallow==3.20.1
gunicorn==21.2.0
python-jose[cryptography]==3.3.0
Werkzeug==2.3.7
psycopg2-binary==2.9.7
pytest==7.4.2
black==23.9.1
flake8==6.1.0
mypy==1.5.1